Move CreateDefaultHandler method to separate class
