+++ Softeq.XToolkit.Remote/Client/DefaultHttpMessageHandlerBuilder.cs
@@ -10,25 +10,25 @@ namespace Softeq.XToolkit.Remote.Client
     public class DefaultHttpMessageHandlerBuilder : HttpMessageHandlerBuilder
     {
         public DefaultHttpMessageHandlerBuilder()
-            : this(CreateDefaultHandler())
+            : this(HttpMessageHandlerProvider.CreateDefaultHandler())
         {
         }
 
         public DefaultHttpMessageHandlerBuilder(HttpMessageHandler? primaryHandler)
         {
             PrimaryHandler = primaryHandler;
+            AdditionalHandlers = new List<DelegatingHandler>();
         }
 
         /// <inheritdoc />
-        public override IList<DelegatingHandler> AdditionalHandlers { get; } = new List<DelegatingHandler>();
+        public override IList<DelegatingHandler> AdditionalHandlers { get; }
 
         /// <inheritdoc />
         public override HttpMessageHandler Build()
         {
             if (PrimaryHandler == null)
             {
-                var message = string.Format("The '{0}' must not be null.", nameof(PrimaryHandler));
-                throw new InvalidOperationException(message);
+                throw new InvalidOperationException($"The '{nameof(PrimaryHandler)}' must not be null.");
             }
 
             return CreateHandlerPipeline(PrimaryHandler, AdditionalHandlers);
